+++ src/main/java/nextstep/ladder/LadderGame.java
@@ -1,5 +1,6 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
